getPlanTranslations instead of usePlanTranslations
