+++ src/domain/license/plans/utils/getPlanTranslations.ts
@@ -1,0 +1,17 @@
+import { TFunction } from 'react-i18next';
+
+interface PlanTranslation {
+  name: string;
+  displayName: string;
+  priceDescription: string;
+  features: string[];
+  disclaimer: string;
+  disabledDisclaimer?: string;
+}
+
+export const getPlanTranslations = (t: TFunction<'translation', undefined>): Record<string, PlanTranslation> => {
+  return t('plansTable.plans', { returnObjects: true }).reduce(
+    (acc, plan: PlanTranslation) => ({ ...acc, [plan.name]: plan }),
+    {}
+  );
+};
